+++ mars/services/task/execution/mars/executor.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 import asyncio
-import itertools
 import logging
 import sys
 from collections import defaultdict
@@ -27,22 +26,20 @@ from .....core.operand import (
     OperandStage,
     ShuffleProxy,
 )
-from .....dataframe.core import DATAFRAME_TYPE, SERIES_TYPE
 from .....lib.aio import alru_cache
 from .....oscar.profiling import (
     ProfilingData,
 )
 from .....resource import Resource
-from .....typing import TileableType, BandType, ChunkType
-from .....utils import Timer, get_chunk_params
-from .....tensor.core import TENSOR_TYPE
+from .....typing import TileableType, BandType
+from .....utils import Timer
 from ....cluster.api import ClusterAPI
 from ....lifecycle.api import LifecycleAPI
-from ....meta.api import MetaAPI, WorkerMetaAPI
+from ....meta.api import MetaAPI
 from ....scheduling import SchedulingAPI
 from ....subtask import Subtask, SubtaskResult, SubtaskStatus, SubtaskGraph
 from ...core import Task
-from ..api import TaskExecutor, ExecutionChunkResult, register_executor_cls
+from ..api import TaskExecutor, register_executor_cls
 from .resource import ResourceEvaluator
 from .stage import TaskStageProcessor
 
@@ -155,6 +152,7 @@ class MarsTaskExecutor(TaskExecutor):
             chunk_graph,
             subtask_graph,
             list(available_bands),
+            tile_context,
             self._scheduling_api,
             self._meta_api,
         )
@@ -164,9 +162,7 @@ class MarsTaskExecutor(TaskExecutor):
         resource_evaluator.evaluate()
         self._stage_processors.append(stage_processor)
         self._cur_stage_processor = stage_processor
-        execution_chunk_results = await stage_processor.run()
-        await self._update_result_meta(execution_chunk_results, tile_context)
-        return execution_chunk_results
+        return await stage_processor.run()
 
     async def __aexit__(self, exc_type, exc_val, exc_tb):
         # clean ups
@@ -459,80 +455,3 @@ class MarsTaskExecutor(TaskExecutor):
         # TODO(Catch-Bull): Pop asyncio.Event when current subtask `set_subtask_result`
         # will never be called
         self._subtask_decref_events[subtask.subtask_id].set()
-
-    async def _update_result_meta(
-        self,
-        execution_chunk_results: List[ExecutionChunkResult],
-        tile_context: Dict[TileableType, TileableType],
-    ):
-        chunk_to_chunk_result = {r.chunk: r for r in execution_chunk_results}
-        update_meta_chunks = chunk_to_chunk_result.keys() - set(
-            itertools.chain.from_iterable(
-                (c.data for c in tiled_tileable.chunks)
-                for tiled_tileable in tile_context.values()
-            )
-        )
-
-        worker_meta_api_to_chunk_delays = defaultdict(dict)
-        for c in update_meta_chunks:
-            address = chunk_to_chunk_result[c].meta["bands"][0][0]
-            meta_api = await WorkerMetaAPI.create(self._session_id, address)
-            call = meta_api.get_chunk_meta.delay(
-                c.key, fields=list(get_chunk_params(c).keys())
-            )
-            worker_meta_api_to_chunk_delays[meta_api][c] = call
-        for tileable in tile_context.values():
-            chunks = [c.data for c in tileable.chunks]
-            for c, params_fields in zip(chunks, self._get_params_fields(tileable)):
-                address = chunk_to_chunk_result[c].meta["bands"][0][0]
-                meta_api = await WorkerMetaAPI.create(self._session_id, address)
-                call = meta_api.get_chunk_meta.delay(c.key, fields=params_fields)
-                worker_meta_api_to_chunk_delays[meta_api][c] = call
-        coros = []
-        for worker_meta_api, chunk_delays in worker_meta_api_to_chunk_delays.items():
-            coros.append(worker_meta_api.get_chunk_meta.batch(*chunk_delays.values()))
-        worker_metas = await asyncio.gather(*coros)
-        for chunk_delays, metas in zip(
-            worker_meta_api_to_chunk_delays.values(), worker_metas
-        ):
-            for c, meta in zip(chunk_delays, metas):
-                chunk_to_chunk_result[c].meta = meta
-
-    @classmethod
-    def _get_params_fields(cls, tileable: TileableType):
-        params_fields = []
-        fields = get_chunk_params(tileable.chunks[0])
-        if isinstance(tileable, DATAFRAME_TYPE):
-            for c in tileable.chunks:
-                cur_fields = set(fields)
-                if c.index[1] > 0:
-                    # skip fetch index_value for i >= 1 on column axis
-                    cur_fields.remove("index_value")
-                if c.index[0] > 0:
-                    # skip fetch dtypes_value for i >= 1 on index axis
-                    cur_fields.remove("dtypes_value")
-                if c.index[0] > 0 and c.index[1] > 0:
-                    # fetch shape only for i == 0 on index or column axis
-                    cur_fields.remove("shape")
-                params_fields.append(list(cur_fields))
-        elif isinstance(tileable, SERIES_TYPE):
-            for c in tileable.chunks:
-                cur_fields = set(fields)
-                if c.index[0] > 0:
-                    # skip fetch name and dtype for i >= 1
-                    cur_fields.remove("name")
-                    cur_fields.remove("dtype")
-                params_fields.append(list(cur_fields))
-        elif isinstance(tileable, TENSOR_TYPE):
-            for i, c in enumerate(tileable.chunks):
-                cur_fields = set(fields)
-                if c.ndim > 1 and all(j > 0 for j in c.index):
-                    cur_fields.remove("shape")
-                if i > 0:
-                    cur_fields.remove("dtype")
-                    cur_fields.remove("order")
-                params_fields.append(list(cur_fields))
-        else:
-            for _ in tileable.chunks:
-                params_fields.append(list(fields))
-        return params_fields
